fix(program): address expansion review nits
Tighten bytecode PC mapper validation, make the expanded-bytecode flag test assert parity, and apply small API cleanups from review. Also ignore Lake build artifacts when embedding zklean templates so workspace clippy is not affected by local .lake directories.

+++ crates/jolt-program/src/image/elf.rs
@@ -105,7 +105,7 @@ fn decode_text_section(
         let first_halfword = u16::from_le_bytes([raw_data[offset], raw_data[offset + 1]]);
         if (first_halfword & 0b11) != 0b11 {
             if first_halfword != 0 {
-                let word = uncompress_rv64_instruction(first_halfword as u32);
+                let word = uncompress_rv64_instruction(first_halfword);
                 let instruction = super::decode::decode_instruction(word, address, true)?;
                 instructions.push(instruction);
             }
